add Layout for pages
