+++ pages/index.tsx
@@ -1,40 +1,18 @@
-import Head from 'next/head'
-import dynamic from 'next/dynamic'
-import styles from '../styles/Home.module.scss'
 import About from '../components/about/About'
-import Header from '../components/header/Header'
 import {serverSideTranslations} from 'next-i18next/serverSideTranslations'
-import Nav from '../components/nav/Nav'
-import {useRouter} from 'next/router'
-
-const DynamicFooter = dynamic(() => import('../components/footer/Footer'))
+import Layout from '../components/layouts/layout'
 
 export default function Home() {
-  const route = useRouter().pathname
-
   return (
-    <div className={styles.container}>
-      <Head>
-        <title>CodepointTKU</title>
-        <meta
-          name="description"
-          content="List of projects, current events and achievements of Codepoint Turku"
-        />
-        <link rel="icon" href="/favicon.ico" />
-      </Head>
-      <Header route={route} />
-      <main className={styles.main}>
-        <About />
-      </main>
-      <DynamicFooter />
-      <Nav route={route} />
-    </div>
+    <Layout>
+      <About />
+    </Layout>
   )
 }
 
 export async function getStaticProps({locale}: any) {
   return {
     props: {...(await serverSideTranslations(locale, ['common', 'footer']))},
-    revalidate: 60
+    revalidate: 60,
   }
 }
